backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic[email]==2.10.3
pydantic-settings==2.6.1
PyJWT==2.10.1
bcrypt==4.2.1
python-multipart==0.0.19
requests==2.32.3
resend==2.10.0
httpx==0.27.2
user-agents==2.2.0
cloudinary==1.41.0
slowapi==0.1.9
jinja2==3.1.4
alembic==1.14.1
sentry-sdk[fastapi]>=2.0.0
